+++ src/extension-test/unit/cursor-doc/paredit-test.ts
@@ -87,6 +87,220 @@ describe('paredit', () => {
             });
         })
 
+        describe('forwardHybridSexpRange', () => {
+            it('Finds end of string', () => {
+                const a = docFromTextNotation('"This |needs to find the end of the string."');
+                const b = docFromTextNotation('"This |needs to find the end of the string.|"');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds newline in multi line string', () => {
+                const a = docFromTextNotation('"This |needs to find the end\n of the string."');
+                const b = docFromTextNotation('"This |needs to find the end|\n of the string."');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds newline in multi line string (Windows)', () => {
+                const a = docFromTextNotation('"This |needs to find the end\r\n of the string."');
+                const b = docFromTextNotation('"This |needs to find the end|\r\n of the string."');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of comment', () => {
+                const a = docFromTextNotation('(a |;; foo\n e)');
+                const b = docFromTextNotation('(a |;; foo|\n e)');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of comment (Windows)', () => {
+                const a = docFromTextNotation('(a |;; foo\r\n e)');
+                const b = docFromTextNotation('(a |;; foo|\r\n e)');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Maintains balanced delimiters 1', () => {
+                const a = docFromTextNotation('(a| b (c\n d) e)');
+                const b = docFromTextNotation('(a| b (c\n d)| e)');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Maintains balanced delimiters 1 (Windows)', () => {
+                const a = docFromTextNotation('(a| b (c\r\n d) e)');
+                const b = docFromTextNotation('(a| b (c\r\n d)| e)');
+                const [start,end] = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                // off by 1 because \r\n is treated as 1 char?
+                expect(actual).toEqual([start, end - 1]);
+            });
+
+            it('Maintains balanced delimiters 2', () => {
+                const a = docFromTextNotation('(aa| (c (e\nf)) g)');
+                const b = docFromTextNotation('(aa| (c (e\nf))|g)');
+                const expected  = textAndSelection(b)[1];
+                const actual =  paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Maintains balanced delimiters 2 (Windows)', () => {
+                const a = docFromTextNotation('(aa| (c (e\r\nf)) g)');
+                const b = docFromTextNotation('(aa| (c (e\r\nf))|g)');
+                const [start,end] = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                // off by 1 because \r\n is treated as 1 char?
+                expect(actual).toEqual([start, end - 1]);
+            });
+
+            it('Maintains balanced delimiters 3', () => {
+                const a = docFromTextNotation('(aa| (  c (e\nf)) g)');
+                const b = docFromTextNotation('(aa| (  c (e\nf))|g)');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Advances past newline when invoked on newline', () => {
+                const a = docFromTextNotation('(a|\n e) g)');
+                const b = docFromTextNotation('(a|\n| e)');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of vectors', () => {
+                const a = docFromTextNotation('[a [b |c d e f] g h]');
+                const b = docFromTextNotation('[a [b |c d e f|] g h]');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of lists', () => {
+                const a = docFromTextNotation('(foo |bar)\n');
+                const b = docFromTextNotation('(foo |bar|)\n');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+
+            it('Finds end of maps', () => {
+                const a = docFromTextNotation('{:a 1 |:b 2 :c 3}');
+                const b = docFromTextNotation('{:a 1 |:b 2 :c 3|}');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of line in multiline maps', () => {
+                const a = docFromTextNotation('{:a 1 |:b 2\n:c 3}');
+                const b = docFromTextNotation('{:a 1 |:b 2|:c 3}');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of expr in multiline maps', () => {
+                const a = docFromTextNotation('{:a 1 |:b (+\n 0\n 2\n) :c 3}');
+                const b = docFromTextNotation('{:a 1 |:b (+\n 0\n 2\n)| :c 3}');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of line in bindings', () => {
+                const a = docFromTextNotation('(let [|a (+ 1 2)\n b (+ 2 3)] (+ a b))');
+                const b = docFromTextNotation('(let [|a (+ 1 2)|\n b (+ 2 3)] (+ a b))');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds end of expr in multiline bindings', () => {
+                const a = docFromTextNotation('(let [|a (+\n 1 \n 2)\n b (+ 2 3)] (+ a b))');
+                const b = docFromTextNotation('(let [|a (+\n 1 \n 2)|\n b (+ 2 3)] (+ a b))');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            });
+
+            it('Finds range in line of tokens', () => {
+                const a = docFromTextNotation(' | 2 "hello" :hello/world\nbye');
+                const b = docFromTextNotation(' | 2 "hello" :hello/world|\nbye');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Finds range in token with form over multiple lines', () => {
+                const a = docFromTextNotation(' | 2 [\n 1 \n]');
+                const b = docFromTextNotation(' | 2 [\n 1 \n]|');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Deals with comments start of line', () => {
+                const a = docFromTextNotation('|;;  hi\n');
+                const b = docFromTextNotation('|;;  hi|\n');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Deals with comments middle of line', () => {
+                const a = docFromTextNotation(';; |hi\n');
+                const b = docFromTextNotation(';; |hi|\n');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Deals with empty lines', () => {
+                const a = docFromTextNotation('|\n');
+                const b = docFromTextNotation('|\n|');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Deals with comments with empty line', () => {
+                const a = docFromTextNotation(';; |\n');
+                const b = docFromTextNotation(';; |\n|');
+                const expected = textAndSelection(b)[1];
+                const actual = paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Does not advance when on closing token type ', () => {
+                const a = docFromTextNotation('(a e|)\n');
+                const b = docFromTextNotation('(a e||)\n');
+                const expected = textAndSelection(b)[1];
+                const actual =  paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+
+            it('Finds the full form after an ignore marker', () => {
+                // https://github.com/BetterThanTomorrow/calva/pull/1293#issuecomment-927123696
+                const a = docFromTextNotation('(comment•  #_|[a b (c d•              e•              f) g]•  :a•)');
+                const b = docFromTextNotation('(comment•  #_|[a b (c d•              e•              f) g]|• :a•)');
+                const expected = textAndSelection(b)[1];
+                const actual =  paredit.forwardHybridSexpRange(a);
+                expect(actual).toEqual(expected);
+            })
+        })
+
         describe('moveToRangeRight', () => {
             it('Places cursor at the right end of the selection', () => {
                 const a = docFromTextNotation('(def |>|foo|>| [vec])');
@@ -844,6 +1058,6 @@ describe('paredit', () => {
                 paredit.addRichComment(a);
                 expect(textAndSelection(a)).toEqual(textAndSelection(b));
             });
-        }) 
+        })
     });
 });